+++ stories/index.js
@@ -1,3 +1,4 @@
 export { default as DataTableStory } from './DataTable';
 export { default as TextFilterStory } from './TextFilter';
 export { default as CheckFilterStory } from './CheckFilter';
+export { default as FilterIndicatorStory } from './FilterIndicator';