+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # React Filter Table
 
+#### v.0.1.3
+* *Feature*: Added Filter Indicator component
+
 #### v.0.1.2
 * *Feature*: Added TextFilter component
+* *Feature*: Added CheckFilter component
 
 #### v.0.1.1
 * *Feature*: Added Storybook